fix: code refactoring and build issues
